Fast Food
A fast-food restaurant and most of the food that you're offering is previously prepared. You need to know if you will have enough food to serve lunch to all of your customers. Write a program that checks the orders' quantity. You also want to know the client with the biggest order for the day, because you want to give him a discount the next time he comes.