Snake Moves
A task to visualize the snake's path in a square form. A snake is represented by a string. The isle is a rectangular matrix of size N x M. A snake starts going down from the top-left corner and slithers its way down. The first cell is filled with the first symbol of the snake, the second cell is filled with the second symbol, etc. The snake is as long as it takes to fill the stairs– if you reach the end of the string representing the snake, start again at the beginning.